requirements.txt
python-nmap>=0.7.1
paramiko>=3.0.0
requests>=2.28.0
click>=8.1.0
rich>=13.0.0
Jinja2>=3.1.0
